# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Itinerary API",
        "description": "API for managing itineraries",
        "contact": {
            "email": "mohamadtalemsi@gmail.com"
        },
        "license": {
            "name": "talemsi abdellah",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/itineraries/all": {
            "get": {
                "summary": "Retrieve a list of itineraries",
                "description": "/**\nRetrieves a list of itineraries.",
                "operationId": "b440b6ee10ae2efcb33142361f43e1ab",
                "responses": {
                    "200": {
                        "description": "List of itineraries retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "itineraries": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/itineraries": {
            "post": {
                "tags": [
                    "Itineraries"
                ],
                "summary": "Create a new itinerary",
                "description": "Store a newly created resource in storage.",
                "operationId": "9ddf03b1ccb2fb24c86fa810f0614b38",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "category",
                                    "duration"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "category": {
                                        "type": "string"
                                    },
                                    "duration": {
                                        "type": "integer"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Itinerary created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Itinerary created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "format": "int64",
                                                    "example": "1"
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Itinerary Title"
                                                },
                                                "category": {
                                                    "type": "string",
                                                    "example": "Category"
                                                },
                                                "duration": {
                                                    "type": "integer",
                                                    "example": "5"
                                                },
                                                "image_url": {
                                                    "type": "string",
                                                    "example": "http://example.com/image.jpg"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-03-25 12:00:00"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-03-25 12:00:00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "title": [
                                                    "The title field is required."
                                                ],
                                                "category": [
                                                    "The category field is required."
                                                ],
                                                "duration": [
                                                    "The duration field is required."
                                                ],
                                                "image": [
                                                    "The image must be an image.",
                                                    "The image may not be greater than 2048 kilobytes."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/itineraries/searchcategory": {
            "get": {
                "tags": [
                    "Itineraries"
                ],
                "summary": "Search itineraries by category",
                "description": "Search itineraries by category.",
                "operationId": "e431c40e9d81de136e121d26a4019244",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Search query for category",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of itineraries matching the category search",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "itineraries": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "format": "int64",
                                                        "example": "1"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Itinerary Title"
                                                    },
                                                    "category": {
                                                        "type": "string",
                                                        "example": "Category"
                                                    },
                                                    "duration": {
                                                        "type": "integer",
                                                        "example": "5"
                                                    },
                                                    "image_url": {
                                                        "type": "string",
                                                        "example": "http://example.com/image.jpg"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-03-25 12:00:00"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-03-25 12:00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}